+++ Assets/Examples/BallMaker.cs
@@ -4,20 +4,36 @@ using UnityEngine;
 using UnityEngine.XR.iOS;
 
 public class BallMaker : MonoBehaviour {
+	private const int DefaultMaxBalls = 100;
 
 	public GameObject ballPrefab;
 	public float createHeight;
+	[Range (1, DefaultMaxBalls)]
+	public int maxBalls = DefaultMaxBalls;
 	private MaterialPropertyBlock props;
+	private List<GameObject> balls = new List<GameObject> ();
 
 	// Use this for initialization
 	void Start () {
+		RepairMaxBalls ();
 		props = new MaterialPropertyBlock ();
 
 	}
 
+	void OnValidate () {
+		RepairMaxBalls ();
+	}
+
 	void CreateBall(Vector3 atPosition)
 	{
+		if (ballPrefab == null) {
+			return;
+		}
+
 		GameObject ballGO = Instantiate (ballPrefab, atPosition, Quaternion.identity);
+		if (ballGO == null) {
+			return;
+		}
 			
 		
 		float r = Random.Range(0.0f, 1.0f);
@@ -27,12 +43,21 @@ public class BallMaker : MonoBehaviour {
 		props.SetColor("_InstanceColor", new Color(r, g, b));
 
 		MeshRenderer renderer = ballGO.GetComponent<MeshRenderer>();
-		renderer.SetPropertyBlock(props);
+		if (renderer != null) {
+			renderer.SetPropertyBlock(props);
+		}
+
+		balls.Add (ballGO);
+		TrimBallsToLimit ();
 
 	}
 
 	// Update is called once per frame
 	void Update () {
+		RepairMaxBalls ();
+		PruneMissingBalls ();
+		TrimBallsToLimit ();
+
 		if (Input.touchCount > 0 )
 		{
 			var touch = Input.GetTouch(0);
@@ -57,6 +82,43 @@ public class BallMaker : MonoBehaviour {
 			}
 		}
 
+	}
+
+	void OnDisable () {
+		ClearBalls ();
+	}
+
+	private void TrimBallsToLimit () {
+		while (balls.Count > maxBalls) {
+			GameObject oldestBall = balls [0];
+			balls.RemoveAt (0);
+			if (oldestBall != null) {
+				Destroy (oldestBall);
+			}
+		}
+	}
+
+	private void ClearBalls () {
+		foreach (GameObject ball in balls) {
+			if (ball != null) {
+				Destroy (ball);
+			}
+		}
+		balls.Clear ();
+	}
+
+	private void PruneMissingBalls () {
+		for (int index = balls.Count - 1; index >= 0; index--) {
+			if (balls [index] == null) {
+				balls.RemoveAt (index);
+			}
+		}
+	}
+
+	private void RepairMaxBalls () {
+		if (maxBalls < 1 || maxBalls > DefaultMaxBalls) {
+			maxBalls = DefaultMaxBalls;
+		}
 	}
 
 }